Uses short for PathSegment. Results for plaintext microbenchmark

**Before**

           Method |       Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
----------------- |-----------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
         Baseline |   181.8 ns |  1.692 ns |  1.413 ns | 5,501,495.5 |   1.00 |     0.00 | 0.0002 |      40 B |
              Dfa |   317.2 ns |  6.213 ns |  7.155 ns | 3,152,332.2 |   1.75 |     0.04 | 0.0010 |     168 B |
 LegacyTreeRouter | 2,451.0 ns | 16.982 ns | 15.054 ns |   407,990.9 |  13.49 |     0.13 |      - |     480 B |
     LegacyRouter | 1,345.2 ns | 14.328 ns | 11.964 ns |   743,363.7 |   7.40 |     0.08 | 0.0019 |     344 B |

**After**

           Method |       Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
----------------- |-----------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
         Baseline |   184.8 ns |  3.759 ns |  3.516 ns | 5,411,169.6 |   1.00 |     0.00 | 0.0002 |      40 B |
              Dfa |   312.5 ns |  2.708 ns |  2.262 ns | 3,199,525.4 |   1.69 |     0.03 | 0.0010 |     168 B |
 LegacyTreeRouter | 2,167.9 ns | 15.722 ns | 14.706 ns |   461,275.4 |  11.73 |     0.22 |      - |     480 B |
     LegacyRouter | 1,476.3 ns | 28.429 ns | 26.593 ns |   677,351.5 |   7.99 |     0.20 | 0.0019 |     344 B |

int16.MaxValue is 32k so that's plenty of space for the URL sizes we allow.