@davidfowl 2018-08-10
Are those results repeatable? This looks like noise