@rynowak 2018-08-10
It might be noise. What I'm trying to do is make the zero-initialization of [this](https://github.com/aspnet/Routing/blob/release/2.2/src/Microsoft.AspNetCore.Routing/Matching/DfaMatcher.cs#L41) less costly